requirements.txt
asgiref==3.6.0
Django==4.1.5
django-filter==22.1
Pillow==9.4.0
python-stdnum==1.18
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
wincertstore==0.2
